+++ src/modules/axpel/screens/Exercise-2.tsx
@@ -1,7 +1,7 @@
 import React, { ReactElement } from 'react'
-import Footer from '../../components/elements/Footer/Footer'
-import Header from '../../components/elements/Header/Header'
-import Map from '../../axpel/components/Map'
+import Footer from '../../../components/elements/Footer/Footer'
+import Header from '../../../components/elements/Header/Header'
+import Map from '../components/Map'
 import Searchbar from '../components/Searchbar'
 import { LoadScript } from '@react-google-maps/api'
 import { Grid, Container } from '@material-ui/core'
